requirements.txt
pipecat-ai
aiohttp
python-dotenv
loguru
piper-tts
onnxruntime
flask
pipecat-ai[local-smart-turn-v3]
boto3
pipecat-ai[daily,runner,webrtc,tracing]
pipecat-ai[aws,openai,deepgram,whisper,xtts]
fastapi
websockets
faster-whisper
opentelemetry-sdk
nltk